JDP220801-24 OrderController implementacja wstępna 2.2 (dodanie ResponseEntity, globalna obsługa wyjątków, dodanie metody updateOrder do serwisu)

+++ src/main/java/com/kodilla/ecommercee/service/DbOrderService.java
@@ -1,13 +1,12 @@
 package com.kodilla.ecommercee.service;
 
 import com.kodilla.ecommercee.domain.Order;
+import com.kodilla.ecommercee.exceptions.OrderNotFoundException;
 import com.kodilla.ecommercee.repository.OrderRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -15,19 +14,32 @@ public class DbOrderService {
 
     private final OrderRepository orderRepository;
 
-    public Optional<Order> findOrderById(Long id) {
-        return orderRepository.findById(id);
+    public Order findOrderById(final Long id) throws OrderNotFoundException {
+        return orderRepository.findById(id).orElseThrow(OrderNotFoundException::new);
     }
 
     public List<Order> findAllOrders() {
         return orderRepository.findAll();
     }
 
-    public Order saveOrder(Order order) {
+    public Order saveOrder(final Order order) {
         return orderRepository.save(order);
     }
 
-    public void deleteOrder(Long id) {
-        orderRepository.deleteById(id);
+    public void deleteOrder(final Long id) throws OrderNotFoundException {
+        if (orderRepository.existsById(id)) {
+            orderRepository.deleteById(id);
+        } else {
+            throw new OrderNotFoundException();
+        }
+    }
+
+    public Order updateOrder(final Order order) throws OrderNotFoundException {
+        if (orderRepository.existsById(order.getId())) {
+            orderRepository.save(order);
+            return order;
+        } else {
+            throw new OrderNotFoundException();
+        }
     }
 }
